fix bugs

+++ src/views/ruang/Index.vue
@@ -4,7 +4,7 @@
       :headers="headers"
       :items="ruangs"
       :sort-by.sync="nama"
-      :search="searc"
+      :search="search"
       class="elevation"
     >
       <template v-slot:top>
@@ -30,51 +30,67 @@
                 <v-btn color="primary" dark class="mb-2" v-on="on">New Item</v-btn>
               </template>
               <v-card>
-                <v-card-title>
-                  <span class="headline">{{formTitle}}</span>
-                </v-card-title>
-                <v-card-text>
-                  <v-container>
-                    <v-row>
-                      <v-col cols="12" sm="6" md="4">
-                        <v-text-field
-                          v-model="editedItem.nama"
-                          label="Nama Ruang">
-                        </v-text-field>
-                      </v-col>
-                      <v-col cols="12" sm="6" md="4">
-                        <v-textarea 
-                          v-model="editedItem.keterangan" label="Keterangan">
-                        </v-textarea>
-                      </v-col>
-                      <v-col cols="12" sm="6" md="4">
-                        <v-text-field 
-                          v-model="editedItem.kapasitas"
-                          label="Kapasitas">
-                        </v-text-field>
-                      </v-col>
-                      <v-col cols="12" sm="6" md="4">
-                        <!-- <v-select
-                          v-model="editedItem.status"
-                          :items="statusItem"
-                          label="Status"
-                        ></v-select> -->
-                        <v-switch
-                          v-model="editedItem.status"
-                          label="Status"
-                        ></v-switch>
-                      </v-col>
-                    </v-row>
-                  </v-container>
-                </v-card-text>
-                <v-card-actions>
-                  <v-spacer></v-spacer>
-                  <v-btn color="blue darken-1" text @click="close()">Close</v-btn>
-                  <v-btn color="blue darken-1" text @click="save()">Save</v-btn>
-                </v-card-actions>
+                <v-form
+                  ref="form"
+                  v-model="formValidation"
+                  lazy-validation
+                >
+                  <v-card-title>
+                    <span class="headline">{{formTitle}}</span>
+                  </v-card-title>
+                  <v-card-text>
+                    <v-container>
+                      <v-row>
+                        <v-col cols="12">
+                          <v-text-field
+                            v-model="editedItem.nama"
+                            label="Nama Ruang"
+                            :rules="inputRules.nama"
+                            required
+                          ></v-text-field>
+                        </v-col>
+                        <v-col cols="12">
+                          <v-textarea 
+                            v-model="editedItem.keterangan" 
+                            label="Keterangan"
+                            :rules="inputRules.keterangan"
+                            required
+                          ></v-textarea>
+                        </v-col>
+                        <v-col cols="12">
+                          <v-text-field 
+                            v-model="editedItem.kapasitas"
+                            label="Kapasitas"
+                            :rules="inputRules.kapasitas"
+                            required
+                          ></v-text-field>
+                        </v-col>
+                        <v-col cols="12">
+                          <!-- <v-select
+                            v-model="editedItem.status"
+                            :items="statusItem"
+                            label="Status"
+                          ></v-select> -->
+                          <v-switch
+                            v-model="editedItem.status"
+                            label="Status"
+                          ></v-switch>
+                        </v-col>
+                      </v-row>
+                    </v-container>
+                  </v-card-text>
+                  <v-card-actions>
+                    <v-spacer></v-spacer>
+                    <v-btn color="blue darken-1" text @click="close()">Close</v-btn>
+                    <v-btn color="blue darken-1" :disabled="!formValidation" text @click="save()">Save</v-btn>
+                  </v-card-actions>
+                </v-form>
               </v-card>
             </v-dialog>
           </v-toolbar>
+      </template>
+      <template v-slot:item.status="{item}">
+        {{(item.status) ? 'Booked' : 'Avaiable'}}
       </template>
       <template v-slot:item.aksi="{item}">
         <v-icon
@@ -100,14 +116,28 @@
 <script>
 export default {
   created(){
-    let uri = 'http://localhost:3000/prodi';
+    let uri = 'http://localhost:3000/ruang';
     this.axios.get(uri)
     .then((res) => {
-      console.log(res);
+      this.ruangs = res.data.data;
     });
   },
   data(){
     return{
+      formValidation:true,
+      inputRules:{
+        nama:[
+          v => !!v || 'nama is required'
+        ],
+        keterangan:[
+          v => (v.length) || 'keterangan is required'
+        ],
+        kapasitas:[
+          v => !!v || 'kapasitas is required',
+          v => (!isNaN(v)) || 'kapasitas berupa angka'
+        ]
+
+      },
       nama:'nama',
       search:'',
       dialog:false,
@@ -116,7 +146,7 @@ export default {
         {text:'Nama', value:'nama'},
         {text:'Keterangan', value:'keterangan'},
         {text:'Kapasitas', value:'kapasitas'},
-        {text:'Status', value:'text_status'},
+        {text:'Status', value:'status'},
         {text:'Aksi', value:'aksi'},
       ],
       editedIndex: -1,
@@ -156,6 +186,7 @@ export default {
   },
   methods:{
     close(){
+      this.$refs.form.resetValidation();
       this.dialog = false;
       this.$nextTick(() => {
         this.editedItem = Object.assign({}, this.defaultItem);
@@ -169,6 +200,7 @@ export default {
           .then((res) => {
             if(res.data.data){
               Object.assign(this.ruangs[this.editedIndex], this.editedItem);
+              this.close();
             }else{
               alert(res.data.error);
             }
@@ -181,31 +213,34 @@ export default {
           .then((res) => {
             if(res.data.data){
               this.ruangs.push(res.data.data);
+              this.close();
             }
           }).catch((err) => {
             alert(err);
           });
       }
-      this.close();
     },
     editItem(item){
       this.editedIndex = this.ruangs.indexOf(item);
-      this.editedItwm = Object.assign({}, item);
+      this.editedItem = Object.assign({}, item);
       this.dialog = true;
     },
     deleteItem(item){
-      let uri = `http://localhost:3000/ruang/${item.id}`;
-      this.axios.delete(uri)
-        .then((res) => {
-          if(res.data.data){
-            let index = this.ruangs.indexOf(item);
-            this.ruangs.splice(index, 1);
-          }else{
-            alert(res.data.error);
-          }
-        }).catch((err) => {
-          alert(err);
-        });
+      let del = confirm('are you sure want to delete this item?');
+      if(del){
+        let uri = `http://localhost:3000/ruang/${item.id}`;
+        this.axios.delete(uri)
+          .then((res) => {
+            if(res.data.data){
+              let index = this.ruangs.indexOf(item);
+              this.ruangs.splice(index, 1);
+            }else{
+              alert(res.data.error);
+            }
+          }).catch((err) => {
+            alert(err);
+          });
+      }
     }
   }
 }
